Fix duplicated types
