+++ plugin-flex-ts-template-v2/src/feature-library/conversation-transfer/custom-components/ParticipantsTab/ParticipantsTab.tsx
@@ -5,8 +5,7 @@ import { ConversationState, styled, Actions } from '@twilio/flex-ui';
 
 import { Participants } from './Participants.tsx/Participants';
 import { InvitedParticipants } from './InvitedParticipants/InvitedParticipants';
-import { ParticipantDetails } from '../../types/ParticipantDetails';
-import { InvitedParticipantDetails } from '../../types/InvitedParticipantDetails';
+import { InvitedParticipantDetails, ParticipantDetails } from '../../../../types/conversations/Participant';
 import { getUpdatedParticipantDetails, getUpdatedInvitedParticipantDetails } from './hooks';
 import {
   CancelChatParticipantInviteActionPayload,
