--- plugin-flex-ts-template-v2/src/feature-library/conversation-transfer/types/InvitedParticipantDetails.ts
@@ -1,13 +1,0 @@
-import { ParticipantInviteType } from './ParticipantInvite';
-
-export interface InvitedParticipantDetails {
-  invitesTaskSid: string;
-  targetSid: string;
-  targetName: string;
-  timestampCreated: Date;
-  inviteTargetType: ParticipantInviteType;
-}
-
-export interface InvitedParticipants {
-  invites: (InvitedParticipantDetails | undefined)[];
-}
