--- plugin-flex-ts-template-v2/src/feature-library/conversation-transfer/types/ParticipantDetails.ts
@@ -1,9 +1,0 @@
-export type ParticipantType = 'agent' | 'worker';
-
-export interface ParticipantDetails {
-  friendlyName: string;
-  participantType: ParticipantType;
-  isMe: boolean;
-  interactionParticipantSid: string;
-  conversationMemberSid: string;
-}
