--- plugin-flex-ts-template-v2/src/feature-library/conversation-transfer/types/ParticipantInvite.ts
@@ -1,18 +1,0 @@
-export type ParticipantInviteType = 'Worker' | 'Queue';
-
-export interface WorkerParticipantInvite {
-  full_name: string;
-  activity_name: string;
-  worker_sid: string;
-  available: boolean;
-}
-
-export interface QueueParticipantInvite {
-  queue_name: string;
-  queue_sid: string;
-}
-
-export interface ParticipantInvite {
-  type: ParticipantInviteType;
-  participant: WorkerParticipantInvite | QueueParticipantInvite | null;
-}
